+++ src/paperless_tesseract/tests/samples/no-text.png
